backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
dnspython>=2.6.0
python-whois>=0.9.0
httpx>=0.27.0
python-dotenv>=1.0.0
